# perf: Add mergeHistogram fast path for single value from