--- a/aggregators/merger.go
+++ b/aggregators/merger.go
@@ -421,26 +421,52 @@ func mergeHistogram(to, from *aggregationpb.HDRHistogram) {
 		return
 	}
 
-	// Single pass O(N + M) merge.
-	var toIdx, fromIdx int
-	for toIdx, fromIdx = 0, 0; toIdx < toLen && fromIdx < fromLen; toIdx++ {
-		v := from.Buckets[fromIdx] - to.Buckets[toIdx]
+	requiredLen := toLen + fromLen
+	for toIdx, fromIdx := 0, 0; toIdx < toLen && fromIdx < fromLen; {
+		v := to.Buckets[toIdx] - from.Buckets[fromIdx]
+		switch {
+		case v == 0:
+			// For every bucket that is common, we need one less bucket in final slice
+			requiredLen--
+			toIdx++
+			fromIdx++
+		case v < 0:
+			toIdx++
+		case v > 0:
+			fromIdx++
+		}
+	}
+
+	toIdx, fromIdx := toLen-1, fromLen-1
+	to.Buckets = slices.Grow(to.Buckets, requiredLen-toLen)[:requiredLen]
+	to.Counts = slices.Grow(to.Counts, requiredLen-toLen)[:requiredLen]
+	for idx := len(to.Buckets) - 1; idx >= 0; idx-- {
+		if fromIdx < 0 {
+			break
+		}
+		if toIdx < 0 {
+			to.Counts[idx] = from.Counts[fromIdx]
+			to.Buckets[idx] = from.Buckets[fromIdx]
+			fromIdx--
+			continue
+		}
+		v := to.Buckets[toIdx] - from.Buckets[fromIdx]
 		switch {
 		case v == 0:
 			to.Counts[toIdx] += from.Counts[fromIdx]
-			fromIdx++
-		case v < 0: // to.Buckets[toIdx] > from.Buckets[fromIdx]
-			to.Buckets[toIdx], from.Buckets[fromIdx] = from.Buckets[fromIdx], to.Buckets[toIdx]
-			to.Counts[toIdx], from.Counts[fromIdx] = from.Counts[fromIdx], to.Counts[toIdx]
+			to.Counts[idx] = to.Counts[toIdx]
+			to.Buckets[idx] = to.Buckets[toIdx]
+			toIdx--
+			fromIdx--
 		case v > 0:
+			to.Counts[idx] = to.Counts[toIdx]
+			to.Buckets[idx] = to.Buckets[toIdx]
+			toIdx--
+		case v < 0:
+			to.Counts[idx] = from.Counts[fromIdx]
+			to.Buckets[idx] = from.Buckets[fromIdx]
+			fromIdx--
 		}
-	}
-	extra := fromLen - fromIdx
-	if extra > 0 {
-		to.Buckets = slices.Grow(to.Buckets, extra)[:toLen+extra]
-		to.Counts = slices.Grow(to.Counts, extra)[:toLen+extra]
-		copy(to.Buckets[toLen:], from.Buckets[fromIdx:])
-		copy(to.Counts[toLen:], from.Counts[fromIdx:])
 	}
 }