--- a/aggregators/merger_test.go
+++ b/aggregators/merger_test.go
@@ -1145,48 +1145,18 @@ func TestMergeHistogram(t *testing.T) {
 		expected *aggregationpb.HDRHistogram
 	}{
 		{
-			name: "from_between_to",
+			name: "non_single_value",
 			to: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000, 3000},
-				Counts:  []int64{1, 3},
+				Buckets: []int32{1, 3, 5, 7, 9},
+				Counts:  []int64{1, 3, 5, 7, 9},
 			},
 			from: &aggregationpb.HDRHistogram{
-				Buckets: []int32{2000},
-				Counts:  []int64{2},
+				Buckets: []int32{2, 4, 5, 8},
+				Counts:  []int64{2, 4, 5, 8},
 			},
 			expected: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000, 2000, 3000},
-				Counts:  []int64{1, 2, 3},
-			},
-		},
-		{
-			name: "to_between_from",
-			to: &aggregationpb.HDRHistogram{
-				Buckets: []int32{2000},
-				Counts:  []int64{2},
-			},
-			from: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000, 3000},
-				Counts:  []int64{1, 3},
-			},
-			expected: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000, 2000, 3000},
-				Counts:  []int64{1, 2, 3},
-			},
-		},
-		{
-			name: "merge_counts",
-			to: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000, 2000},
-				Counts:  []int64{1, 2},
-			},
-			from: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000, 2000},
-				Counts:  []int64{1, 2},
-			},
-			expected: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000, 2000},
-				Counts:  []int64{2, 4},
+				Buckets: []int32{1, 2, 3, 4, 5, 7, 8, 9},
+				Counts:  []int64{1, 2, 3, 4, 10, 7, 8, 9},
 			},
 		},
 		{
@@ -1196,18 +1166,18 @@ func TestMergeHistogram(t *testing.T) {
 				Counts:  []int64{},
 			},
 			from: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000},
+				Buckets: []int32{1},
 				Counts:  []int64{1},
 			},
 			expected: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000},
+				Buckets: []int32{1},
 				Counts:  []int64{1},
 			},
 		},
 		{
 			name: "empty_from",
 			to: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000},
+				Buckets: []int32{1},
 				Counts:  []int64{1},
 			},
 			from: &aggregationpb.HDRHistogram{
@@ -1215,23 +1185,38 @@ func TestMergeHistogram(t *testing.T) {
 				Counts:  []int64{},
 			},
 			expected: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000},
+				Buckets: []int32{1},
 				Counts:  []int64{1},
 			},
 		},
 		{
-			name: "single_from_exist",
+			name: "single_value_exist",
 			to: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000, 2000, 3000},
+				Buckets: []int32{1, 2, 3},
 				Counts:  []int64{1, 2, 3},
 			},
 			from: &aggregationpb.HDRHistogram{
-				Buckets: []int32{3000},
+				Buckets: []int32{3},
 				Counts:  []int64{4},
 			},
 			expected: &aggregationpb.HDRHistogram{
-				Buckets: []int32{1000, 2000, 3000},
+				Buckets: []int32{1, 2, 3},
 				Counts:  []int64{1, 2, 7},
+			},
+		},
+		{
+			name: "single_value_not_exist",
+			to: &aggregationpb.HDRHistogram{
+				Buckets: []int32{1, 2, 4},
+				Counts:  []int64{1, 2, 4},
+			},
+			from: &aggregationpb.HDRHistogram{
+				Buckets: []int32{3},
+				Counts:  []int64{3},
+			},
+			expected: &aggregationpb.HDRHistogram{
+				Buckets: []int32{1, 2, 3, 4},
+				Counts:  []int64{1, 2, 3, 4},
 			},
 		},
 	} {