Optimize mergeHistogram for single value in `from` histogram, since it will be the majority case.

Alternative to #66. 

Also attached benchstat of BenchmarkMergeHistogram. It is a microbenchmark that measures time to merge single value `from` histogram to `to` histogram of size `size` and whether that the single value already exists in `to.Buckets`. It is not super well written and possibly only useful for this time. Therefore I've not included it in the change, but only attaching it in this PR description.

benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                                             │ bench-out/merge-histogram-fastpath-8-before │ bench-out/merge-histogram-fastpath-8-after │
                                             │                   sec/op                    │       sec/op         vs base               │
AggregateCombinedMetrics-16                                                    1.890µ ± 6%         1.878µ ±   5%        ~ (p=0.699 n=6)
AggregateBatchSerial-16                                                        4.454µ ± 3%         4.403µ ± 144%        ~ (p=0.818 n=6)
AggregateBatchParallel-16                                                      5.087µ ± 5%         5.036µ ±   3%        ~ (p=0.589 n=6)
CombinedMetricsEncoding-16                                                     870.7n ± 3%         922.9n ±   4%   +6.00% (p=0.004 n=6)
CombinedMetricsToBatch-16                                                      60.60µ ± 4%         64.23µ ±   4%   +6.00% (p=0.004 n=6)
EventToCombinedMetrics-16                                                      394.3n ± 3%         387.4n ±   3%        ~ (p=0.225 n=6)
NDJSONSerial/go-2.0.0.ndjson-16                                                17.07m ± 4%         14.99m ±   4%  -12.19% (p=0.002 n=6)
NDJSONSerial/nodejs-3.29.0.ndjson-16                                           9.765m ± 3%         8.874m ±   5%   -9.12% (p=0.002 n=6)
NDJSONSerial/python-6.7.2-labels.ndjson-16                                     38.70m ± 4%         39.76m ±   4%        ~ (p=0.093 n=6)
NDJSONSerial/python-6.7.2.ndjson-16                                            23.12m ± 4%         23.01m ±   3%        ~ (p=0.589 n=6)
NDJSONSerial/ruby-4.5.0.ndjson-16                                              14.14m ± 5%         13.17m ±   8%   -6.85% (p=0.015 n=6)
NDJSONParallel/go-2.0.0.ndjson-16                                              16.80m ± 4%         15.09m ±   4%  -10.17% (p=0.002 n=6)
NDJSONParallel/nodejs-3.29.0.ndjson-16                                         9.664m ± 4%         8.792m ±   5%   -9.02% (p=0.002 n=6)
NDJSONParallel/python-6.7.2-labels.ndjson-16                                   40.45m ± 5%         39.31m ±   3%        ~ (p=0.240 n=6)
NDJSONParallel/python-6.7.2.ndjson-16                                          24.08m ± 9%         24.01m ±   5%        ~ (p=0.485 n=6)
NDJSONParallel/ruby-4.5.0.ndjson-16                                            14.27m ± 4%         13.20m ±   6%   -7.45% (p=0.009 n=6)
geomean                                                                        710.7µ              688.2µ          -3.16%

                                             │ bench-out/merge-histogram-fastpath-8-before │ bench-out/merge-histogram-fastpath-8-after │
                                             │                    B/op                     │        B/op         vs base                │
AggregateCombinedMetrics-16                                                 1.104Ki ± 1%           1.107Ki ± 0%       ~ (p=0.387 n=6)
AggregateBatchSerial-16                                                     1.373Ki ± 2%           1.349Ki ± 2%  -1.74% (p=0.037 n=6)
AggregateBatchParallel-16                                                   1.367Ki ± 1%           1.373Ki ± 1%       ~ (p=0.372 n=6)
CombinedMetricsEncoding-16                                                    0.000 ± 0%             0.000 ± 0%       ~ (p=1.000 n=6) ¹
CombinedMetricsToBatch-16                                                   3.982Ki ± 0%           3.981Ki ± 0%       ~ (p=0.550 n=6)
EventToCombinedMetrics-16                                                     0.000 ± 0%             0.000 ± 0%       ~ (p=1.000 n=6) ¹
NDJSONSerial/go-2.0.0.ndjson-16                                             9.235Mi ± 0%           9.249Mi ± 0%       ~ (p=0.310 n=6)
NDJSONSerial/nodejs-3.29.0.ndjson-16                                        5.570Mi ± 1%           5.555Mi ± 1%       ~ (p=1.000 n=6)
NDJSONSerial/python-6.7.2-labels.ndjson-16                                  29.95Mi ± 1%           30.02Mi ± 1%       ~ (p=0.699 n=6)
NDJSONSerial/python-6.7.2.ndjson-16                                         15.39Mi ± 1%           15.35Mi ± 0%       ~ (p=0.310 n=6)
NDJSONSerial/ruby-4.5.0.ndjson-16                                           8.249Mi ± 0%           8.253Mi ± 0%       ~ (p=0.485 n=6)
NDJSONParallel/go-2.0.0.ndjson-16                                           9.232Mi ± 0%           9.230Mi ± 1%       ~ (p=0.937 n=6)
NDJSONParallel/nodejs-3.29.0.ndjson-16                                      5.558Mi ± 1%           5.566Mi ± 1%       ~ (p=0.818 n=6)
NDJSONParallel/python-6.7.2-labels.ndjson-16                                30.06Mi ± 1%           29.92Mi ± 1%       ~ (p=0.180 n=6)
NDJSONParallel/python-6.7.2.ndjson-16                                       15.30Mi ± 1%           15.37Mi ± 0%       ~ (p=0.240 n=6)
NDJSONParallel/ruby-4.5.0.ndjson-16                                         8.251Mi ± 0%           8.253Mi ± 0%       ~ (p=0.818 n=6)
geomean                                                                                  ²                       -0.06%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                             │ bench-out/merge-histogram-fastpath-8-before │ bench-out/merge-histogram-fastpath-8-after │
                                             │                  allocs/op                  │     allocs/op       vs base                │
AggregateCombinedMetrics-16                                                   17.00 ± 0%             17.00 ± 0%       ~ (p=1.000 n=6) ¹
AggregateBatchSerial-16                                                       13.00 ± 0%             13.00 ± 8%       ~ (p=1.000 n=6)
AggregateBatchParallel-16                                                     13.00 ± 8%             13.00 ± 8%       ~ (p=1.000 n=6)
CombinedMetricsEncoding-16                                                    0.000 ± 0%             0.000 ± 0%       ~ (p=1.000 n=6) ¹
CombinedMetricsToBatch-16                                                     95.00 ± 0%             95.00 ± 0%       ~ (p=1.000 n=6) ¹
EventToCombinedMetrics-16                                                     0.000 ± 0%             0.000 ± 0%       ~ (p=1.000 n=6) ¹
NDJSONSerial/go-2.0.0.ndjson-16                                              114.6k ± 1%            114.0k ± 0%       ~ (p=0.485 n=6)
NDJSONSerial/nodejs-3.29.0.ndjson-16                                         63.15k ± 2%            63.50k ± 1%       ~ (p=0.485 n=6)
NDJSONSerial/python-6.7.2-labels.ndjson-16                                   326.3k ± 1%            326.3k ± 1%       ~ (p=0.937 n=6)
NDJSONSerial/python-6.7.2.ndjson-16                                          163.4k ± 0%            162.6k ± 2%       ~ (p=0.180 n=6)
NDJSONSerial/ruby-4.5.0.ndjson-16                                            98.69k ± 0%            99.01k ± 1%       ~ (p=0.065 n=6)
NDJSONParallel/go-2.0.0.ndjson-16                                            113.8k ± 1%            114.0k ± 1%       ~ (p=0.818 n=6)
NDJSONParallel/nodejs-3.29.0.ndjson-16                                       62.87k ± 1%            63.96k ± 1%  +1.72% (p=0.009 n=6)
NDJSONParallel/python-6.7.2-labels.ndjson-16                                 326.3k ± 1%            328.1k ± 1%       ~ (p=0.394 n=6)
NDJSONParallel/python-6.7.2.ndjson-16                                        162.3k ± 1%            162.9k ± 1%       ~ (p=0.589 n=6)
NDJSONParallel/ruby-4.5.0.ndjson-16                                          98.77k ± 0%            98.86k ± 0%       ~ (p=0.589 n=6)
geomean                                                                                  ²                       +0.18%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

MergeHistogram benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                               │ bench-out/merge-histogram-main │  bench-out/merge-histogram-bench-9  │
                               │             sec/op             │    sec/op     vs base               │
MergeHistogram/100-false-16                        148.45n ± 1%   93.07n ±  7%  -37.31% (p=0.002 n=6)
MergeHistogram/100-true-16                         174.10n ± 3%   75.12n ±  2%  -56.85% (p=0.002 n=6)
MergeHistogram/1000-false-16                       1113.0n ± 2%   437.8n ± 10%  -60.67% (p=0.002 n=6)
MergeHistogram/1000-true-16                        1373.0n ± 4%   162.2n ±  4%  -88.19% (p=0.002 n=6)
MergeHistogram/10000-false-16                      12.181µ ± 1%   4.998µ ±  3%  -58.97% (p=0.002 n=6)
MergeHistogram/10000-true-16                       14.437µ ± 2%   2.009µ ±  2%  -86.08% (p=0.002 n=6)
MergeHistogram/100000-false-16                     148.13µ ± 3%   75.81µ ±  4%  -48.82% (p=0.002 n=6)
MergeHistogram/100000-true-16                      171.57µ ± 2%   46.72µ ±  5%  -72.77% (p=0.002 n=6)
geomean                                             4.527µ        1.432µ        -68.38%

                               │ bench-out/merge-histogram-main │ bench-out/merge-histogram-bench-9  │
                               │              B/op              │    B/op     vs base                │
MergeHistogram/100-false-16                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/100-true-16                         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/1000-false-16                       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/1000-true-16                        0.000 ±  ?     0.000 ± 0%       ~ (p=1.000 n=6)
MergeHistogram/10000-false-16                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/10000-true-16                       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/100000-false-16                     0.000 ± 0%     0.000 ±  ?       ~ (p=1.000 n=6)
MergeHistogram/100000-true-16                      0.000 ±  ?     0.000 ± 0%       ~ (p=1.000 n=6)
geomean                                                       ²               +0.00%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                               │ bench-out/merge-histogram-main │ bench-out/merge-histogram-bench-9  │
                               │           allocs/op            │ allocs/op   vs base                │
MergeHistogram/100-false-16                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/100-true-16                         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/1000-false-16                       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/1000-true-16                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/10000-false-16                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/10000-true-16                       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/100000-false-16                     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
MergeHistogram/100000-true-16                      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                       ²               +0.00%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

Benchmark function:
```go
func BenchmarkMergeHistogram(b *testing.B) {
	factor := 10
	for _, tc := range []struct {
		size int
	}{
		{
			size: 100,
		},
		{
			size: 1000,
		},
		{
			size: 10000,
		},
		{
			size: 100000,
		},
	} {
		for _, hit := range []bool{false, true} {
			b.Run(fmt.Sprintf("%d-%v", tc.size, hit), func(b *testing.B) {
				baseBuckets := make([]int32, tc.size)
				baseCounts := make([]int64, tc.size)

				for i := 0; i < tc.size; i++ {
					r := rand.Int31n(int32(factor))
					if i == 0 {
						baseBuckets[i] = r
					} else {
						baseBuckets[i] = baseBuckets[i-1] + r
					}
					baseCounts[i] = int64(r)
				}

				buckets := make([]int32, tc.size+1)
				counts := make([]int64, tc.size+1)

				to := aggregationpb.HDRHistogram{
					Buckets: buckets,
					Counts:  counts,
				}
				from := aggregationpb.HDRHistogram{
					Buckets: []int32{0},
					Counts:  []int64{1},
				}
				b.ResetTimer()

				for i := 0; i < b.N; i++ {
					copy(counts[:], baseCounts[:])
					copy(buckets[:], baseBuckets[:])
					to.Buckets = buckets[:tc.size]
					to.Counts = counts[:tc.size]
					if hit {
						from.Buckets[0] = to.Buckets[rand.Intn(tc.size)]
					} else {
						from.Buckets[0] = rand.Int31n(int32(tc.size * factor))
					}
					mergeHistogram(&to, &from)
				}
			})
		}
	}
}
```